# docs/swagger.yaml
definitions:
...
  controllers.CommentResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      metadata:
        $ref: '#/definitions/controllers.PaginationMetadata'
    type: object
...
  controllers.MessageResponse:
    properties:
      message:
        type: string
    type: object
  controllers.PaginationMetadata:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  controllers.Request:
    properties:
      content:
        type: string
    type: object
  controllers.SingleCommentResponse:
    properties:
      comment:
        $ref: '#/definitions/models.Comment'
      parent_comment:
        $ref: '#/definitions/models.Comment'
    type: object
...
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      parent_id:
        type: integer
      post_id:
        type: integer
      replies:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Like:
    properties:
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Post:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      comments_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      i_liked:
        type: boolean
      id:
        type: integer
      image_urls:
        items:
          type: string
        type: array
      likes:
        items:
          $ref: '#/definitions/models.Like'
        type: array
      likes_count:
        type: integer
      post_type:
        type: string
      shares_count:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      views_count:
        type: integer
    type: object
  models.User:
    properties:
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      followers:
        description: Self-referential many-to-many for followers and following.
        items:
          $ref: '#/definitions/models.User'
        type: array
      following:
        items:
          $ref: '#/definitions/models.User'
        type: array
      id:
        type: integer
      name:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      profile_picture:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Allows the comment’s author to delete it
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - Comments
    get:
      consumes:
      - application/json
      description: Get a comment with its replies and user information
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SingleCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      summary: Get a single comment by ID
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Allows the comment’s author to update it
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Edit a comment
      tags:
      - Comments
  /comments/{id}/replies:
    post:
      consumes:
      - application/json
      description: Allows a user to reply to an existing comment. The reply is also
        a comment, with a ParentID linking it to the original comment.
      parameters:
      - description: Parent Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reply content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a reply to a comment
      tags:
      - Comments
...
  /posts/{id}/comments:
    get:
      consumes:
      - application/json
      description: Get paginated comments for a post, including replies
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Comments per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      summary: Get comments for a specific post
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Allows a user to comment on a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a comment to a post
      tags:
      - Comments
...
swagger: "2.0"